NPE in clear temp folder solved

+++ src/main/java/br/com/margel/updater/StepClearTempFiles.java
@@ -17,7 +17,11 @@ class StepClearTempFiles extends UpdaterStep{
 		try {
 			tempFolder = updater.getConfig().getTempDownloadFolderPath();
 			LOGGER.debug("Clear temp files -> "+tempFolder);
-			for(File f : new File(tempFolder).listFiles()) {
+			File[] tempFiles = new File(tempFolder).listFiles();
+			if(tempFiles==null) {
+				return;
+			}
+			for(File f : tempFiles) {
 				Files.delete(f.toPath());
 				LOGGER.debug("Delete -> "+f);
 			}
